chore: wallet 의존이슈로 entity 수정, currentcoin 타입 수정

+++ src/main/java/com/picktartup/startup/dto/StartupElasticsearch.java
@@ -48,7 +48,7 @@ public class StartupElasticsearch {
     private Integer goalCoin;
 
     @Field(name = "current_coin", type = FieldType.Integer)
-    private Integer currentCoin;
+    private Double currentCoin;
 
     @Field(name = "expected_return", type = FieldType.Double)
     private Double expectedReturn;
